requirements.txt
PyQt6==6.5.2
paddlepaddle==2.5.1
paddleocr==2.7.0
Pillow==10.0.0
numpy==1.24.3
opencv-python==4.6.0.66 
keyboard==0.13.5